refactor(upload > views > toolbar): rename and refactor to class WatermarkButton
