--- src/admin/upload/views/ToolbarView/WatermarkButton.js
@@ -1,89 +1,0 @@
-import AddWatermarkButton, {
-  generateAndSaveMarkedImage,
-} from "../../components/AddWatermarkButton";
-
-const WatermarkButton = AddWatermarkButton.extend({
-  initialize(props) {
-    this.controller = props.controller;
-    this.selection = props.selection;
-    this.isSingle = false;
-
-    AddWatermarkButton.prototype.initialize.apply(this);
-
-    this._addStatesEvents();
-  },
-
-  render() {
-    AddWatermarkButton.prototype.render.apply(this);
-
-    this.el.classList.add("button-primary");
-    this._hide();
-
-    return this;
-  },
-
-  _showResult(result) {
-    if (result) {
-      if (0 === result.length) {
-        alert("no image selected");
-      } else {
-        const validResult = result.filter((image) => image);
-
-        alert(`${validResult.length} / ${result.length} images added.`);
-      }
-
-      this.selection.reset();
-    } else {
-      alert("Encountered some issues. Image has not been created.");
-    }
-  },
-
-  _addStatesEvents() {
-    this.controller.on("select:activate", () => {
-      this._show();
-
-      if (0 === this.selection.length) {
-        this._disable();
-      }
-
-      this.controller.on("selection:toggle", () => {
-        if (0 === this.selection.length) {
-          this._disable();
-        } else if (this._isDisabled()) {
-          this._enable();
-        }
-      });
-    });
-
-    this.controller.on("select:deactivate", () => {
-      this._hide();
-    });
-
-    this.selection.on("reset", (e) => {
-      this._disable();
-    });
-
-    this.on("saveResult", (data) => {
-      this._showResult(data);
-    });
-  },
-
-  _hide() {
-    this.el.classList.add("hidden");
-  },
-  _show() {
-    this.el.classList.remove("hidden");
-  },
-
-  _isDisabled() {
-    return this.el.classList.contains("disabled");
-  },
-  _disable() {
-    this.el.classList.add("disabled");
-  },
-  _enable() {
-    this.el.classList.remove("disabled");
-  },
-});
-
-export default WatermarkButton;
